requirements.txt
openai
llama-index-core
llama-index-llms-openai
llama-index-embeddings-huggingface
python-dotenv
sentence-transformers

